requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
opencv-python
tqdm
transformers>=4.36.0
accelerate
pillow
scikit-learn